backend/requirements.txt
# Forensic CTF Platform - Backend Dependencies
# Python 3.12+

# Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
argon2-cffi==23.1.0
passlib==1.7.4

# Storage (S3-compatible)
minio==7.2.3

# HTTP Client (for health checks, etc.)
httpx==0.26.0

# Utilities
python-multipart==0.0.9
python-dotenv==1.0.1

# Development
pytest==7.4.4
pytest-asyncio==0.23.4
pytest-cov==4.1.0
black==24.1.1
ruff==0.2.0
mypy==1.8.0
